Check all schemas in search_path, not just the topmost one.
The previous version of this code implicitly assumed
that the topmost schema in the search_path (usually
'public') is the only one that needs to be searched.
This leads to issues with multi-schema deployments.

Postgres has a built-in function to check all
entries in the search_path (see
https://www.postgresql.org/docs/current/functions-info.html#FUNCTIONS-INFO-SESSION),
current_schemas(include_implicit boolean), and this
is what we are using to solve this issue.